dell-dock: Use gsize for result length calculation
The result_length variable was guint8 but computed as length + 1 where
length is gsize. This would truncate if the requested length exceeds
254 bytes.

(cherry picked from commit ff6d9ee873e0b62bebba8a55004b568c72c61420)

+++ plugins/dell-dock/fu-dell-dock-ec.c
@@ -260,7 +260,7 @@ fu_dell_dock_ec_read(FuDellDockEc *self, guint32 cmd, gsize length, GBytes **byt
 	FuDevice *proxy;
 	/* The first byte of result data will be the size of the return,
 	hide this from callers */
-	guint8 result_length = length + 1;
+	gsize result_length = length + 1;
 	g_autoptr(GBytes) bytes_local = NULL;
 	const guint8 *result;
